pyproject.toml
[project]
name = "pipelines"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.8.18"
dependencies = [
    "apache-airflow-providers-postgres",
    "pandas==2.0.3",
    "psycopg2-binary",
    "splink>=4.0.11",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "ruff>=0.15.2",
]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".tox",
    ".venv",
    "__pypackages__",
    "build",
    "dist",
    "node_modules",
    "site-packages",
]

[tool.ruff.lint]
select = ["E", "F", "I"]
fixable = ["ALL"]

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
